requirements.txt
annotated-types==0.7.0
attrs==19.3.0
Automat==0.8.0
bleach==6.1.0
blinker==1.4
blis==0.7.11
catalogue==2.0.10
certifi==2024.12.14
chardet==3.0.4
click==8.1.8
cloud-init==24.4
cloudpathlib==0.20.0
colorama==0.4.3
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dbus-python==1.2.16
distro==1.4.0
distro-info==0.23+ubuntu1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.3
filelock==3.16.1
fonttools==4.55.5
fsspec==2024.12.0
httplib2==0.14.0
huggingface-hub==0.27.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.8.1
importlib_resources==6.4.5
incremental==16.10.1
Jinja2==2.10.1
joblib==1.4.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
kaggle==1.6.17
keyring==18.0.1
kiwisolver==1.4.7
langcodes==3.4.1
language-selector==0.1
language_data==1.3.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==1.1.0
matplotlib==3.7.5
mdurl==0.1.2
more-itertools==4.2.0
mpmath==1.3.0
murmurhash==1.0.10
netifaces==0.10.4
networkx==3.1
nltk==3.9.1
numpy==1.24.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
packaging==24.2
panda==0.3.1
pandas==2.0.3
pexpect==4.6.0
pillow==10.4.0
preshed==3.0.9
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.19.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-slugify==8.0.4
pytz==2024.2
PyYAML==5.3.1
regex==2024.11.6
requests==2.22.0
requests-unixsocket==0.2.0
rich==13.9.4
safetensors==0.5.2
SecretStorage==2.3.1
service-identity==18.1.0
shellingham==1.5.4
simplejson==3.16.0
six==1.14.0
smart-open==7.1.0
sos==4.7.2
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
ssh-import-id==5.10
sympy==1.13.3
systemd-python==234
text-unidecode==1.3
thinc==8.2.4
tokenizers==0.20.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
transformers==4.46.3
triton==3.0.0
Twisted==18.9.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
wrapt==1.17.2
zipp==3.20.2
zope.interface==4.7.1
